+++ CHANGELOG.md
@@ -2,9 +2,11 @@
 
 This changelog records changes to stable releases since 1.50.2. "TBA" changes here may be available in the [nightly release](https://github.com/microsoft/vscode-js-debug/#nightly-extension) before they're in stable. Note that the minor version (`v1.X.0`) corresponds to the VS Code version js-debug is shipped in, but the patch version (`v1.50.X`) is not meaningful.
 
-## TBA
+## v1.54.0 - 2021-02-04
 
 - fix: allow copying values from watch expressions ([ref](https://github.com/microsoft/vscode/issues/115049))
+- fix: reuse debug terminals when running npm scripts, when possible
+- refactor: move script lens functionality into built-in npm extension
 
 ## v1.53.0 - 2021-01-25